[Refactor] 통합 및 config 클래스 위치 변경
Member에 있던 Config 클래스들 패키지 변경 및 Swagger-ui 적용

+++ src/main/java/com/woof/api/cart/model/dto/CartDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class CartDto {
     Long idx;
     Long productCeoIdx;
-    Integer brandIdx;
+    Long brandIdx;
     String name;
     String filename;
 }
